requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncstdlib==3.12.3
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.0
fastapi==0.110.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
jsonalias==0.1.1
MarkupSafe==2.1.5
nodeenv==1.8.0
packaging==24.0
prisma==0.13.1
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
sniffio==1.3.1
solana==0.33.0
solders==0.21.0
starlette==0.37.2
StrEnum==0.4.15
tomlkit==0.12.4
typing_extensions==4.11.0
uvicorn==0.29.0
websockets==11.0.3
